似乎可以用的版本

+++ libraries/chain/include/eosio/chain/fixed_bytes.hpp
@@ -93,7 +93,7 @@ namespace eosio {
          }
 
       public:
-         typedef uint128_t word_t;
+         typedef uint64_t word_t;
 
          /**
           * Get number of words contained in this fixed_bytes object. A word is defined to be 16 bytes in size
